Rename BaseGoalSelect to BaseGoalWidget
