+++ src/goals/DefaultGoal/BaseGoalScreen/tests/BaseGoalScreen.test.tsx
@@ -7,7 +7,7 @@ import { CreateCharInv } from "../../../CreateCharInv/CreateCharInv";
 
 it("renders without crashing", () => {
   const div = document.createElement("div");
-  let user: User = new User("", "", "");
+  let user: User = new User("TestUser", "TestUsername", "password");
   let goal: Goal = new CreateCharInv([]);
   goal.user = user;
   ReactDOM.render(<BaseGoalScreen goal={goal} />, div);
