+++ src/goals/DefaultGoal/BaseGoalWidget/BaseGoalWidget.tsx
