+++ src/goals/MergeDupGoal/index.tsx
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 
 import { StoreState } from "../../types";
-import BaseGoalSelect from "../DefaultGoal/BaseGoalSelect/BaseGoalSelect";
+import BaseGoalSelect from "../DefaultGoal/BaseGoalWidget/BaseGoalWidget";
 import BaseGoalScreen from "../DefaultGoal/BaseGoalScreen/BaseGoalScreen";
 
 //placeholder interfacees
